Bug 693259 (pref): removeCommand should ignore not-found
On the one hand it could be useful to know if you're not removing a
command when you thought you were, as a result of a typo, on the other
hand, it's a pain if you have to check before you remove in case it's
not there (or provoke an expected exception).
Silent removal ignore seems to be the general default in these sitations.

This arose because in splitting out 'pref list' we were going to have a
startup config flag, which toggled use of the 'pref list' command (which
is too slow when embedded in firefox)