Bug 760793 (nodenull): Protect against null inputs
As for Bug 760791 (tiltopen) for boolean, array and selection types.
It is unlikely that these fixes will be required in the real world because
the blank values for boolean and array are not null (they are false and
empty array respectively) and selection types are commonly 1 from n.
However it's worth being general in our fixes.