Bug 761268 (helpstring): Swap functions for properties
i.e. getHeading() -> heading and getMatchingCommands() -> matchingCommands
This style is better because it means we're more likely to be able to use
allowEval:false, and it makes it easier for the properties to depend on
each other. Also we're about to add more properties, so now is a good time
to reduce use of functions.