Bug 720972 (optmenu): Refactor template data generation
Previously we used the completer instance as the data for the template,
however this didn't make much sense since the set of things that were
important to templates and the set of things important to being a completer
were disjoint sets, and using properties (the old way) meant either
inefficient code that re-calculated properties, or cached values (what we
did) which caused subtle sequencing issues. Now we just have a simple
(if a little large) function to generate the template data.