Bug 720972 (optmenu): Add _toSummaryJson
_toSummaryJson alows us to see at a glance what the state of the command
line it. This is useful partly in debugging because it saves digging down
and could be useful in testing to check the state of a large number of
things at the same time, however we will need to ensure that it's stable
before using it in testing too much.
This patch also changes a demo command that we use for debugging. This
change does not reach Firefox.

REBASE: make summaryJson a property

tweak command