case-773254: Cache lower case versions
match is the input that we're trying to match, in lower case, and we
add a _gcliLowerName property to each option as a lower case version of
the name.

The options are provided by the commands or by gcli itself, so i think
it's safe to add a _gcliLowerName property without causing damage.