options-775031: Add '[options]' to hints
We add '[options]' when there is at least one remaining named parameter,
which is not hidden and which doesn't already have a value associated with
it (i.e. the argument type is not 'BlankArgument').
Also we need to add an nbsp before the '[options]' if there isn't already
a space at the end of the command line.
Also update the many tests affected by this change, and add some new ones.