boolgroup-773565: Named param proposals need spaces
Previously when there were predictions in the arguments to named parameters
we were presenting them without a space. For example if | represents the
cursor we would do this:
>> copy --source|readme.txt
Rather than:
>> copy --source| readme.txt
(The readme.txt is a different color to indicate it's a completion)

The fix involves checking for a NamedArgument without a space at the end of
the prefix and then adding a space to the beginning of the proposal.

We also remove the [options] marker whenever we have a tab proposal, which
is clearer in general, and rename optionsRemaining to addOptionsMarker
since the calculation now involves more than just 'are there options
remaining'.