namedhint-770830: Better hints for named parameters
This is related to my fixes for Bug 773565, and included there for
simplicity.

Previously, we didn't do much with:
>> git commit --message
But we wanted to tell the user they needed to type a message. The previous
commit in this branch added tab completion for named parameters when there
was a prediction, but for things like strings and numbers, we can't make
any useful predictions, so we do the next best thing and tell them what
they need to type using the type name.

The substance of these changes is in the first 2 hunks of completer.js.
The final hunk of change to completer.js looks big but really it's just
the addition of an if statement, which increases the indent of a number of
lines. We're saying if we added a prompt for a string or number above then
don't make any other prompts untill they've completed this task.