boolgroup-773465: Consistent error handling
We setup a Command and set of Parameters when addCommand is called, and
do some basic sanity checking on the data passed in. Previously we were
inconsistent about reporting through console.error or throwing. Experience
has shown that particularly with the noisy Firefox console, that errors
can get lost. So now we throw for all errors.

This changeset also aligns continuation indents with start of string.
Previously continuation indents were 2, 4 or 8 spaces.