boolgroup-773465: Allow strings in emptyParameters
Helpers.check() used to require an array for emptyParameters, but that's
unnecessarily verbose. So now we allow strings, and I threw a regex at the
unit tests to convert the existing arrays to strings.

The code in Helpers is updated in 2 places: check() and status(). status
is the older version of check which has not been purged yet.

The Array version has been kept even though the tests in GCLI don't use it
any more because very similar versions of check() and status() exist in
Firefox (in head.js for the command line), and I'd like to keep the code
in sync as much as possible.