boolgroup-773565: Provide hints for named args
Don't stop presenting hints just because the current argument is a named
argument.

Also make a minor clarification to some documentation.

## About this direction

I've gone both ways on providing hints other arguments while editing
named arguments, and I'm still not 100% sure, but this, I think is the
best route for now.

For example, what should be the hints for the following input, given that
the 'git tag' command requires a tag name and can optionally take a
message [1]:
>> git tag --message foo|

Option 1: Show required params, and options as normal:
>> git tag --message foo|
                          <tag> [options]

Options 2: Show the hints only when we're clearly done with the named arg:
>> git tag --message foo|

>> git tag --message foo |
                          <tag> [options]

For now we're doing option 1.

For the alternative, see boolgroup-773565-experiment.

[1] In reality 'git tag' uses a '-m' prefix for the message, but we're
    showing it here as '--message' to fit the way GCLI currently works.