boolgroup-773565: Tweak/fix Argument.beget()
Arguments are what the user types into the command line. They are designed
to be immutable because a Conversion links an Argument to the value that it
represents, and it can't do that if the Argument changes underneith it.

So when we want to change an Argument (for example completion) then we need
to create a new Argument based on the old one. Hence the beget() function.

beget used to take 2 parameters, the new text and an options object. But
there were 2 problems:
- What if you didn't want to change the text (completion sometimes just
  wants to add a space to the prefix/suffix for example)
- The implementation was terrible in that it did different things just if
  there was an empty options object.

Solution: merge the replText parameter into the options object. Now it is
easier to extend to allow us to alter the prefix/suffix too.

This is a particular issue because MergedArguments were being messed up by
the completion process so we want to create a smart MergedArgument.beget()