boolgroup-773565: Don't use exports alias
Previously we aliased 'exports' so when we moved code from one module to
another, we wouldn't need to redirect calls to functions in the original
module. While a worthy goal, it's often confusing, and in this case not
relevant because there isn't any of that kind of intra-module function
calling.