boolgroup-773565: Move incr/decr to Requisition
Minor refactor: We're debugging completion, part of which is making
Requistion.setAssignment() update the underlying arg list properly. If we
make the increment and decrement process be part of Requisition (rather
than Assignment) then we can get them to use Requistion.setAssignment()
too.

In general, the events that GCLI uses to update various parts of the UI
are too complex. This is, in part, due to Assignments being adjustable by
the UI directly, so the Assignments need to inform the Requisition that
they have changed, so we get an event storm. This helps make Requisition
into a gatekeeper for all change.