boolgroup-773565: TrueNamedArgument needs an Argument
As previously noted Arguments have 2 uses, one is really a Token, i.e. a
digestable lump. The other is something that has been understood, and
assigned to a parameter. TrueNamedArgument bridged both worlds. This
stops that. Now it's something that can be assigned, not a token, so there
is always an underlying 'token' (called Argument until we get around to
finishing off that refactor)