boolgroup-773565: Upgrade Argument.beget()
Previously there was a bug in Argument.beget, in that it would always make
instances of Argument, not whatever the subclass was called. So we add an
option to specify a subclass to create. This means that all subclasses
need to be able to construct themselves from a prefix/text/suffix trio.

So this change:
- Adds options.type and documents its use
- Makes the NamedArgument ctor able to use prefix/text/suffix
- Adds TrueNamedArgument.beget and NamedArgument.beget
- Upgrades ScriptArgument.beget to use options.type