boolgroup-773565: setAssignment should update all args
Previously Requisition.setAssignment just looked for a single argument
(more properly 'token') when it was updating what the user typed. When
NamedArguments were being completed, this broken. so we now use the array
of arguments (tokens) provided to do the update.

This exposed a bug in the implementation of NamedArgument.getArgs where
it was returning the valueArg even in the case were the user had not typed
one.