boolgroup-773565: Calculate the 'next' param properly
In a couple of places we were assuming that the next parameter was just
the parameter defined 'after' this one. This is OK if all of your params
are positional, but when you're doing named parameters, then we might
enter the paremters out of order, so we define a new function called
Requisition._getFirstBlankPositionalAssignment() and use this in place of
(paramIndex + 1) in a couple of places.

This fixes a bug with completion, so we also update a test to check that
we get this right.