boolgroup-773565: Extract addSpace function
The final thing that we need to fix is the need to add a space in the
'already-complete' case, in addition to the current case. So it makes
sense to extract an addSpace function to handle both.

In addition there is some code that I don't understand, but which is
required. I raised bug 779443 to sort it out.