boolgroup-773565: TAB should add space
No predictions generally means we shouldn't change anything on TAB, but
TAB has the connotation of 'next thing' and when we're at the end of
a thing that implies that we should add a space.
So call _addSpace() in this extra case.

Also add tests for this case, and the cases were this should not fire.