boolgroup-773565: Fix test harness
Yikes! Discovered that helpers was testing "if (x)" which was not running
tests when the thing to check for was empty string. The check should be
"if (x != null)".

The bad news is that this simple mistake was made many times.

The good news is that this only affected one test, and that was a case of
the test being wrong rather than being an undetected bug.