boolgroup-773565: Move getPredictionAt
Conversion had a getPredictionAt function. Moving this to Assignment will
allow us to exclude predictions when we in the name part of a named
argument (although that is not part of this changeset).

This also defaults the index to zero, which seems sensible, and fixes some
cut and paste documentation in Conversion.constrainPredictionIndex()