boolgroup-773565: Upgrade Argument.beget
We previously assumed that you would only need to add spaces to the start
of a prefix or end of a suffix, however the split nature of a named
argument means we also need to add a space to the end of a prefix. So we
add a prefixPostSpace option.

Also NamedArgument used to create valueArg replacements whether or not they
were needed. This blocks the fix that we're currently working on so we
only add a valueArg replacement when there is one already or we need one
to hold added text.