boolgroup-773565: Make named arg completions more normal
Alter Assignment.getPredictionAt to not show predictions when we're in the
name part of a named argument, and use getPredictionAt rather than
getPredictions().length to decide how to complete. This means we handle
predictions for named arguments just like other arguments.

Also add a space if TAB is pressed we are in the name part of an
assignment to match the behaviour when we press TAB at the end of a
command.

This changeset also tidies up a few tests by moving some tests into a new
testNoTab function, and removes a blank testBlank function.