boolgroup-773565: Make all hints visible
Earlier in this branch we tried an experiment with not showing some of the
hints when we had directTabText for named arguments. The goal of allowing the
user to 'concentrate on the thing in hand' is a good one, but deciding when
to simplify is complex and often debatable. So this changeset makes all hints
available always.

This change also breaks some tests, they will be fixed by the very next commit.
While this is bisect unsafe, it is my estimation that this breakage will cost
less than the time to rebase between the 2 changesets.