boolgroup-773565: Tests and Completer updates
The previous commit broke some tests around spaces in the completer. This
fixes that.

The problem was that adding space prefixes is complex. Much better to use
space suffixes and not worry about trailing spaces. So this changeset
simplifies completer.js to use space suffixes.

The effect of this is to break a huge number of tests because we need to move
spaces around. So I took this opportunity to combine directTabText,
emptyParameters, and arrowTabText into a single test called 'hints'.