nohelp-773227: Simplify Requisition._addSpace()
Previously we we made adding spaces complex because getCurrentAssignment
didn't correctly predict the next assignment. Those problems have been
fixed, so we can remove much of the 'next assignment' logic in addSpace.
This also fixes the bug where we create a new empty argument where one
is not needed.

This change also adds a number of tests to check that we have fixed the
problem of the help command (and other things with optional arguments) not
working when they are tabbed into.