nohelp-773227: Show tooltip on invalid help
This change to inputter.js is the heart of the fix for this bug - Call
focusManager.setError(true) if the menu system didn't want to use the
return to activate some menu action.

This requires us get the menu system to tell us if there was some menu
activatation happening.