newmutate-739947: setConversion -> setAssignment
Alter setAssignment to take a third parameter which decides if we need
to do the processing to update the args array (which we need to do if
this is not part of an update call).

Also update it to accept either an arg (in which case convert it), or a
conversion (in which case use it directly) or null (in which case use
the blank value).

Also we have 2 versions of the setAssignment function, one external,
which does not expose the skipArgUpdate parameter.

Now all updates to the assignments can go through the setAssignment
function, which means we may be able to control events somewhat more.