newpromise-708984: Use Add-on/Q like promise
Promise is now a copy of toolkit/addon-sdk/promise/core.js, and
Requisition has code to handle both the new way and the old way. Also
pref_list, which depended on the old way, is updated to use defer().

Documentation is updated

Commands that use createPromise() have not been updated in this commit
partly to prove that this system is backwards compatible.

Signed-off-by: Joe Walker <jwalker@mozilla.com>