asyncout-773313: Fix test runner
The test runner was being somewhat overzealous with its calling of
callbacks. Part of the fix is storing the callback so there are fewer
of them. This also moves the asynchronaity to Test and removing it from
the Suite. It needs to be in Test to satisfy assert.checkCalled, and so
we don't really need it in Suite any more.

Signed-off-by: Joe Walker <jwalker@mozilla.com>