asynctypes-685526: Make getCompleterTemplateData async
The changes to completer.js look more significant than they really are
we just introduce three new promisedXXX variables, and do all the
calculations in Q.resolve(prediction) passing the promises of the 3
promises in the return data.

getCompleterTemplateData is used in 2 places, completer.update and
the test helpers. completer.update does not need to change at all
because domtemplate is awesome.

The test helpers are affected in 3 places, the hints function which
has to resolve the 3 promised values, the debug outputted which remains
synchronous by inserting the string "<promise>" when it gets an
unresolved value, and the asserts themselves which use test.checkCalled
to defer the test and automagically include the results in the correct
test.

Signed-off-by: Joe Walker <jwalker@mozilla.com>