asynctypes-685526: Make Requisition.complete async
The changes to testKeyboard mostly involve the indent caused by adding
Q.resolve.

Inputter gains a _completed promise that tracks if there are any
outstanding completions. This is something of an experiment.

There are some unknown consequences of this change with slow to finish
asynchronous types. It is likely that there will be some problems that
we need to work around in the future.

Signed-off-by: Joe Walker <jwalker@mozilla.com>