testconcat-707492: Create 'exports' alias
In short: s/exports/mockCommands/g.
In general we've been reducing the number of exports aliases, because
although it's handy when moving code from one module to another to not
have the functions change name from exports.foo to blah.foo, that's not
a regular occurrence, and it creates complexity. However in this case
where we're importing code into Firefox mochitests where everything is
in the same global, it's much simpler.

Signed-off-by: Joe Walker <jwalker@mozilla.com>