testconcat-707492: s/test/assert/g & fix testResource
Moving test files into firefox JS land brings up a name clash over
'test', where in mochitest 'test' is the name of the initial test
function, and here is is what we call the assert module by default.

It's normal with GCLI to call imported modules by their name, so we are
renaming the import of the assert module to assert, which solves both
problems.

Also a previous changeset in this series removed all of the fakeDom and
firefox guards from the tests. The resource test module still fails
because we're not running the tests in the environment that it expects
so the failing tests are re-skipped.

Signed-off-by: Joe Walker <jwalker@mozilla.com>