testconcat-707492: remove options from exec
For some reason the version of helpers in firefox was one step ahead
of the version of helpers in GCLI in getting rid of the redundant
'options' parameter. This brings GCLI up to date with the firefox way,
needed now because the GCLI tests are now running against the firefox
helpers implementation.

Signed-off-by: Joe Walker <jwalker@mozilla.com>