menuclick-788215: Convert setAssignment to use options
We had setAssignment and _setAssignment, but the distinction (internal
vs. external) wasn't helpful, so we're merging them.
We need to add another customization - to duplicate arg padding, and
we don't want to do that as another parameter, so we are moving to an
options syntax, and removing the negative on argUpdate while we are at
it.

Signed-off-by: Joe Walker <jwalker@mozilla.com>