asynctypes-685526: Add 'predictions' testing
This allows us to check what the menu should display. It is based on a
call to getPredictions rather than the actual menu (which might be a
better test, but can be left for another day)

This makes _createDebugCheck asynchronous so we also adjust the F2
action when in debug mode.

Signed-off-by: Joe Walker <jwalker@mozilla.com>