nodeinout-796921: Add enter/leave/change support to node type
This makes NodeType and NodeListType support the onEnter, onLeave and
onChange events to call into a revised host.Highlighter class to
highlight the selected nodes as they are typed.

Something in this change breaks the tests being run in PhantomJS, but
they still work fine in Chrome and Firefox, so I'm assuming the problem
is with Phantom, and not running the test there.

Signed-off-by: Joe Walker <jwalker@mozilla.com>