noxhtml-768399: Use innerHTML before XHTML alternatives
This change is taken from a patch by Saran <ksk.3393@gmail.com> to
gcli.jsm in mozilla-central. See this attachment for more details:
https://bug768399.bugzilla.mozilla.org/attachment.cgi?id=667062

Use of HTML (i.e. not XHTML) was failing when embedded in HTML documents
where it would otherwise be valid. This reverses the logic to try
innerHTML before createContextualFragment()