spacecomplete-797016: Auto wrap completed args with spaces
Previously if you pressed TAB to complete an arg which contained a space
you'd get the completion but without the quotes. There was code in
Argument.beget() to handle this case but due to a missing negative it
was never called. Also we need to add to NamedArgument some code to
handle the added quotes.

We also needed a special case for commands because we don't want to add
quotes around sub-commands.

Signed-off-by: Joe Walker <jwalker@mozilla.com>