asynctypes-685526: Allow getBlank to be synchronous
Previously getBlank was implemented basically using this.parse('') but
when parse is asynchronous we're either going to have to make getBlank
asynchronous, or implement it another way. Getting the blank value of a
type feels like something that we should always be able to do
synchronously, so this change implements it that way in all cases but
for selection which we can change later.

Signed-off-by: Joe Walker <jwalker@mozilla.com>