asynctypes-685526: Test/bug fixes
ParamType.lookup can't be sure that there will be a command, so it
should guard against this and return an empty set.
testTypes should provide a requisition to all types in the typeSpec,
and while we are at it 'typeSpec' should be called 'typeSpec' and not
'options' which was confusing when the test options got involved.

Signed-off-by: Joe Walker <jwalker@mozilla.com>