asynctypes-685526: Rename DeferredType to DelegateType
The significant issue is that defer() is used in 2 similar contexts one
to do with promises, and one to do with delayed type resolution. I am
renaming the thing i have control over.

This does not rename the 'defer()' method that is part of the
DeferredType contract. That will come in a subsequent commit.

Signed-off-by: Joe Walker <jwalker@mozilla.com>