asynctypes-685526: Make getLookup() async
getLookup and findPredicitons now always return a promise.
increment and decrement will not work with async types with this change.
This change also adds a 'tsslow' mock command which works in trivial
cases. (All synchronous types work as before)
In addition I've added my TODO notes on the outstanding work before
this branch is totally complete.

Signed-off-by: Joe Walker <jwalker@mozilla.com>