ffsync-800346: sync completer.html
The firefox version of completer.html embedded in gcli.jsm is different
to the one in git. I suspect a mistake while creating a patch at some
point. The difference here is use of text nodes vs value attributes.
The difference is text wrapping [1] and the value attribute is possibly
slightly more correct although perhaps not enough to cause us to want to
fix the inconsistency in the markup member.

[1] https://developer.mozilla.org/en-US/docs/XUL/label

Signed-off-by: Joe Walker <jwalker@mozilla.com>