leak-795988: Remove the global commandOutputManager
Sharing a commandOutputManager wasn't the common case, and it caused
confusion, so we make the default to be one commandOutputManager per
requisition, and allow configuration to share them.

This means that Requisition adds a 3rd parameter (a custom COM) which
defaults to a new COM if null. This is then used by tests and Display
implementations to configure a COM. Focus and OutputTerminal then use
the COM in their Requisition rather than having an explicit COM of
their own.

Signed-off-by: Joe Walker <jwalker@mozilla.com>