date-773271: This is a working prototype for a GCLI date type
This works in a demo command named "gcli onedate".
It supports the increment/decrement/parse/stringify operations, but not
min/max: I don't think they make sense for a date. Increments steps are
specifiable, and the unit is currently days. For example, with a step
of 5, 'Tue Jan 15 2013 15:46:29 GMT+0100 (CET)' becomes
'Sun Jan 20 2013 15:46:29 GMT+0100 (CET)'.

Issues:
* testDate.js has not been updated
* There is no integration with the cookie command yet
* MDN says that n should be between 1 and 31 for Date.setDate(n), but
  it seems to work with any number on Firefox and Chrome, and I "abuse"
  this behavior.

Signed-off-by: Quentin Pradet <quentin.pradet@gmail.com>