asynctypes-685526: Mock commands should return strings
They did return lumps of data, originally with some marginal testing
value, but we don't have converters yet, and having something that can
be displayed is now more useful to testing.
The low number of tests that need updating as a result of this change
(i.e. 1) is a testament to how not-useful the lump of data was.

Signed-off-by: Joe Walker <jwalker@mozilla.com>