asynctypes-685526: Add 'exec' to test suite
Previously all testing had to be done by a separate exec function which
made synchronisation hard. This change allows for an exec property to
audit, and handles the promises correctly.

We expect to remove helpers.exec as soon as we have converted the tests
to use helpers.audit in it's place.

Signed-off-by: Joe Walker <jwalker@mozilla.com>