asynctypes-685526: Add forceAsync and assertArgsAssigned
This adds 2 testing updates:

Requisition._assertArgsAssigned which is there to catch a bug which only
shows up with async testing. Somewhere in complete, we are dropping the
assignment from an arg. It could well have something to do with the
TODO line.

Also we're adding forceAsync to selections. Once we're done with the
fix above then we can move onto setting forceAsync to true and making
fixes there.

Signed-off-by: Joe Walker <jwalker@mozilla.com>