asynctypes-685526: Fix issues with complete/addSpace
Previously addSpace assumed that it could force synchronous behaviour
which was a handy cheat, but this couldn't be sustained.
The onTextChange event was firing too early, not giving time for the
async parts to complete; now it is delayed properly.
The extra promise action going on in the complete function meant that we
needed to swap from a 'promise' variable to an 'outstanding' array.

Signed-off-by: Joe Walker <jwalker@mozilla.com>