asynctypes-685526: Enhance test generation
If we're setting non-EOL cursor position, then name the test better and
fix a bug where the completion promise was being dropped.
Also output a template for exec. We're expecting generated tests to be
edited, and it's better to provide something to be deleted rather than
force the user to remember how to do exec.

Signed-off-by: Joe Walker <jwalker@mozilla.com>