asynctypes-685526: Use promises in test harness
This slashes a ton of complex code by removing checkCalled. From now on
a test is done when its promise completes so tests are directly in
control of when they're done. This fixes an long standing bug where when
we flipped the forceAsync flag to true, we got results coming in when we
thought the test had completed, which was because function wrapped in
checkCalled could be called several times.

Tests that wish to complete asynchronously will now need to return a
promise rather than call checkCalled. For most tests this is an easy
job of doing "return helpers.audit([…]);" rather than just calling
"helpers.audit".

The Firefox test integration will need an update, but we're saving that
until later.

Signed-off-by: Joe Walker <jwalker@mozilla.com>