asynctypes-685526: Deprecate exec with string
We can't do exec with a typed string synchronously but this exposed
functionality can work in some cases, so we are deprecating it and will
remove it in the future.

The majority of the changes in this commit are to rename the input
object to 'options' to prioritise its new favoured use, and to update
the documentation.

Signed-off-by: Joe Walker <jwalker@mozilla.com>