asynctypes-685526: Allow async attribute values
Previously <label value="${somePromise}"/> would fail. This update does
the minimum required in that it requires the whole attribute to be a
single promise. It's possible to imagine a solution where we wait until
several promises are resolved before joining resolved values with other
strings, but this isn't that.

This update also removes the export of the deprecated Templater class.
This is important because I'd like to have state be stack based rather
than class based. The Templater class has been deprecated for long
enough that I'm fairly sure it's not being used any more.

Signed-off-by: Joe Walker <jwalker@mozilla.com>