asynctypes-685526: Rewind - rebase
Having tests be 'option' free was good in that it felt simple, but in reality
it just moved the complexity elsewhere. In particular we don't want to hamper
Firefix testing with the setup demands of the test suite here. So we're putting
options back in, with the goal of having mock command be setup by the things
that need them, and helpers needing no setup at all.

Signed-off-by: Joe Walker <jwalker@mozilla.com>