travis-841337: Add Output/promise, fix toString
Output (i.e. what you get from an exec) needs events because the result
of an exec could evolve (e.g. "tail -f") but we also want a simple
'are we done' thing. This adds that so Output now has a then function
which acts like a promise.

Also we fix toString which should have used this.data in place of output
which must have been a cut and paste error. And remove the console.error from
update() to allow errors to propagate.

Signed-off-by: Joe Walker <jwalker@mozilla.com>