travis-841337: Don't include data in Output.resolve
Change 79a1a96 (just 5 commits above) added a promise to the Output
object, which resolves when the execution finishes with the output of
the command. This sounds like a good idea, but in practice it's
confusing:
- The requisition exec resolves to an output object, but when you're
  chaining promises it's easy to forget what you're waiting on
- The data often isn't useful by itself - you're likely to need to know
  the type of the data, or if it's an error, etc. So you're likely to
  need the output object anyway.

You could make an argument for resolve(this), but I'd want to have a
play first, and using undefined is safe - we can expose something else
instead, but we can't take away.

Signed-off-by: Joe Walker <jwalker@mozilla.com>