converters-657595: Better error handling for promises
Previously we were using console.error as the handler for rejected
promises, but this is broken on chrome (console functions not bound) and
it doesn't show stack traces on node, so we introduce a custom error
handler to fix this.

Also completer was handling promises badly, by just logging on error
rather than passing the errors on to the promises that it creates. So
we also update the error handling there.

Signed-off-by: Joe Walker <jwalker@mozilla.com>