exec-729514: Split execution/conversion contexts
Previously there was a single 'context' object that essentially provided
a wrapper for a Requisition to allow us to evolve a Requisition and
hope to have a degree of backwards compatibility and safety. With
remoting and splitting execution from formatting there could be 2
different requisitions, so we split the execution functions from the
conversion functions.

To allow backwards compatibility we have a 'legacy' switch to make the
execution context act as it did previously.

Signed-off-by: Joe Walker <jwalker@mozilla.com>

Conflicts:
	lib/gcli/cli.js
	lib/server/index.js