remotegcli-839862: Update connect command
Previously we were constrained to have only one connection at a time
and we didn't have any route to adding differing types of connection.
By extracting a Connection object, we can easily connect more than once
and by having a Request object we begin to separate the connection
protocol from the rest of GCLI.

Also add a canon.removeProxyCommands function to match the
canon.addProxyCommands that we added earlier.
We also created a new 'connection' type that completes the named
connections.

Signed-off-by: Joe Walker <jwalker@mozilla.com>