types-767589: Allow object literal type definitions
Previously Types all inherited (eventually) from 'Type' or more directly
from SelectionType or DelegateType. They had to do this because various
parts of the system used 'instanceof' to check what they were.

This makes exposing types hard because all new types must first import
types from GCLI, and must then jump through JS inheritance hoops before
registering a new type.

So this change arms 'addType' with the ability to register a new type
using an object literal that contains a 'parent' property which points
to another type whose properties should be used as a 'manual prototype',
that is to say we'll create a new object copy in the parent properties
then copy in the provided properties.

In future commits we'll avoid using instanceof in GCLI code and upgrade
a couple of types to the new system.

We're continuing to allow old-style types in the short term for 2
reasons:
* It's work to remove them, and external things may still need them
* We have some concerns that it might slow the system down as a result
  of confusing the type inference system in JS engines.

Any existing slowdown is in the order of 0.5ms per keypress [1], however
which sounds low, put it has to be put against the overall average time
to process a keypress of around 5ms.

[1] This isn't very scientific though.

Signed-off-by: Joe Walker <jwalker@mozilla.com>