refactor-864370: Hack modules not the help command
GCLI in node-js uses the standard help command, however because we're
outputting to text there is a different template. Previously in order
to get this to be used we'd alter the help module, however with this
change we just define the template that we want help to require.

Changes to code that goes into Firefox:
- The help module no-longer exports hooks to alter the templates
- A new viewToString converter

Signed-off-by: Joe Walker <jwalker@mozilla.com>