exec-754315: Update host.exec() function
Previously the code didn't work at all. This adds a NodeJS version of
host and uses the same pre-define trick to make it work in Node over the
web version that we used for the help hack earlier.

This also updates to web version of host.exec to use XHR to call to the
web server to make the change. And the mozilla version to do nothing.

Signed-off-by: Joe Walker <jwalker@mozilla.com>