refactor-864370: Tighten up conversion types
Add a number of converters from number, boolean, undefined and null to
dom, which could help with commands like 'pref set' that have no return
value.

Complain if the value that we get in the output object is just 'object'
which means that it wasn't specified anywhere.

Have intro.js output a typed object rather than leaving it to be
guessed by something.

Also remove a redundant check:{} from testCanon.js

Signed-off-by: Joe Walker <jwalker@mozilla.com>