Simple querySelectorAll command-770156
Simple querySelectorAll (`qsa`) command added to gcli.
The command takes a string parameter which will be run as a css selector.
Matching nodes will be displayed.
Nodes can then be clicked. When clicked, a unique CSS selector will be
calculated for the given node and `qsa` will be run again with that selector.

The command simply uses document.querySelectorAll in its exec method.
The view is done via a converter.

Signed-off-by: Patrick Brosset <patrickbrosset@gmail.com>