refactor-864370: Output.onChange -> output.promise
Output already had a promise, but it tried to use an onChange function
too in-case of an evolving output. Current thinking is that we should
have long-running commands emit a separate stream object rather than
having Output be both a promise and a stream.

This change removes the onChange event and replaces it with a promise.

Signed-off-by: Joe Walker <jwalker@mozilla.com>