tokenize-786661: Param order: assert.is(actual, expected);
Unit tests in GCLI have changed. Originally assert.is() was called
equals() and the order was equals(expected, actual) because it seemed to
make sense to have the expected thing first, however if the function
name is 'is' then you have to put the 'actual' parameter first. When we
changed from 'equals' to 'is' (to match mochitest) we didn't swap the
orders of the params (but added new asserts the new way). This corrects
that and fixes the orders of the params.

Signed-off-by: Joe Walker <jwalker@mozilla.com>