allowblank-866680: Legally blank parameter hints
Previously for "cookie --path ''" we were still showing the hint for the
path even though we'd clearly entered an empty string. This was because
the test for whether or not we'd entered data was whether the text
of the argument was empty, now we test to see if there is an argument
which means that "" (i.e. no text but valid argument) no longer shows
the hint.

Signed-off-by: Joe Walker <jwalker@mozilla.com>