date-773271: Use custom toISOString()
The problem with toISOString is that it displays the time in UTC, but
we make decisions on time display using local time. This effectively
creates a new version of toISOString which uses local time.

The comments include some discussion of the best separator character
between the date and the time. Feedback welcome.

Signed-off-by: Joe Walker <jwalker@mozilla.com>