remotetest-874821: Prioritize args over command in testing
In helpers.audit() we allow the 'arg' check to check the command itself,
but what if the command has a parameter called 'command' (for example,
an 'exec' command)? We now default to using the parameter because
checking the command value is less useful.

Also we don't need to create a new node in order to convert output.

Signed-off-by: Joe Walker <jwalker@mozilla.com>